Adding support for large objects larger than 2MB.
Limit is now 255MB of object data in a single value.

+++ RazorDB/Key.cs
@@ -157,7 +157,7 @@ namespace RazorDB {
         public static KeyEx Empty {
             get { return new KeyEx(new byte[0], 0); }
         }
-        public KeyEx WithSequence(byte seqNum) {
+        public KeyEx WithSequence(int seqNum) {
             return new KeyEx(KeyBytes, seqNum);
         }
